only return tasks belonging to logged in user

+++ task_list.php
@@ -1,18 +1,19 @@
 <?php
   require 'inc/bootstrap.php';
   isAuthorized();
+  $user = findUserById($session->get('auth_user_id'));
 
   $pageTitle = "Task List | Time Tracker";
   $page = "tasks";
 
   $filter = request()->get('filter');
   if ($filter=='all') {
-    $tasks = getTasks();
+    $tasks = getTasks($user['id']);
   } elseif ($filter=='complete') {
-    $tasks = getCompleteTasks();
+    $tasks = getCompleteTasks($user['id']);
   } else {
     $filter = 'incomplete';
-    $tasks = getIncompleteTasks();
+    $tasks = getIncompleteTasks($user['id']);
   }
 
   include 'inc/header.php';
@@ -38,27 +39,31 @@
         <li<?php if ($filter=='all') echo ' class="on"'; ?>><a href="task_list.php?filter=all">All</a></li>
       </ul>
       <table class="items">
-        <tr><th>Status</th><th>Title</th><th>Action</th></tr>
         <?php
-          foreach ($tasks as $item) {
-            echo "<tr><td>";
-            echo "<input type='checkbox' onChange='javascript:location=\"inc/actions_tasks.php?action=status&task_id=".$item['id'];
-            if (!empty($filter)) {
-              echo "&filter=$filter";
+          if (!empty($tasks) && $tasks != null) {
+            echo "<tr><th>Status</th><th>Title</th><th>Action</th></tr>";
+            foreach ($tasks as $item) {
+              echo "<tr><td>";
+              echo "<input type='checkbox' onChange='javascript:location=\"inc/actions_tasks.php?action=status&task_id=".$item['id'];
+              if (!empty($filter)) {
+                echo "&filter=$filter";
+              }
+              if ($item['status']) {
+                echo "&status=0\"' checked";
+              } else {
+                echo "&status=1\"' ";
+              }
+              echo " />";
+              echo "</td><td width='90%'>";
+              echo "<a href='task.php?id=" . $item['id'] . "'>" . $item['task'] . "</a>";
+              echo "</td><td>";
+              echo "<a href='inc/actions_tasks.php?action=delete&task_id=".$item['id'];
+              echo "' onclick=\"return confirm('Are you sure you want to delete this task?');\"";
+              echo "'>Delete</a>";
+              echo "</td></tr>\n";
             }
-            if ($item['status']) {
-              echo "&status=0\"' checked";
-            } else {
-            echo "&status=1\"' ";
-            }
-            echo " />";
-            echo "</td><td width='90%'>";
-            echo "<a href='task.php?id=" . $item['id'] . "'>" . $item['task'] . "</a>";
-            echo "</td><td>";
-            echo "<a href='inc/actions_tasks.php?action=delete&task_id=".$item['id'];
-            echo "' onclick=\"return confirm('Are you sure you want to delete this task?');\"";
-            echo "'>Delete</a>";
-            echo "</td></tr>\n";
+          } else {
+            echo "You do not have any $filter tasks at this time.";
           }
         ?>
       </table>
